Split up GraphicsManager file into individual class files.
