+++ InteractiveCfd_Core/Graphics/GraphicsManager.cpp
@@ -1,4 +1,6 @@
 #include "GraphicsManager.h"
+#include "ShaderManager.h"
+#include "CudaLbm.h"
 #include "Shader.h"
 #include "Panel.h"
 #include "Layout.h"
@@ -14,629 +16,10 @@
 #undef min
 #undef max
 
-
-CudaLbm::CudaLbm()
-{
-    m_domain = new Domain;
-    m_isPaused = false;
-    m_timeStepsPerFrame = TIMESTEPS_PER_FRAME / 2;
-}
-
-CudaLbm::CudaLbm(const int maxX, const int maxY)
-{
-    m_maxX = maxX;
-    m_maxY = maxY;
-}
-
-Domain* CudaLbm::GetDomain()
-{
-    return m_domain;
-}
-
-float* CudaLbm::GetFA()
-{
-    return m_fA_d;
-}
-
-float* CudaLbm::GetFB()
-{
-    return m_fB_d;
-}
-
-int* CudaLbm::GetImage()
-{
-    return m_Im_d;
-}
-
-float* CudaLbm::GetFloorTemp()
-{
-    return m_FloorTemp_d;
-}
-
-Obstruction* CudaLbm::GetDeviceObst()
-{
-    return m_obst_d;
-}
-
-Obstruction* CudaLbm::GetHostObst()
-{
-    return &m_obst_h[0];
-}
-
-float CudaLbm::GetInletVelocity()
-{
-    return m_inletVelocity;
-}
-
-float CudaLbm::GetOmega()
-{
-    return m_omega;
-}
-
-void CudaLbm::SetInletVelocity(const float velocity)
-{
-    m_inletVelocity = velocity;
-}
-
-void CudaLbm::SetOmega(const float omega)
-{
-    m_omega = omega;
-}
-
-void CudaLbm::SetPausedState(const bool isPaused)
-{
-    m_isPaused = isPaused;
-}
-
-bool CudaLbm::IsPaused()
-{
-    return m_isPaused;
-}
-
-int CudaLbm::GetTimeStepsPerFrame()
-{
-    return m_timeStepsPerFrame;
-}
-
-void CudaLbm::SetTimeStepsPerFrame(const int timeSteps)
-{
-    m_timeStepsPerFrame = timeSteps;
-}
-
-
-
-void CudaLbm::AllocateDeviceMemory()
-{
-    size_t memsize_lbm, memsize_int, memsize_float, memsize_inputs;
-
-    int domainSize = ceil(MAX_XDIM / BLOCKSIZEX)*BLOCKSIZEX*ceil(MAX_YDIM / BLOCKSIZEY)*BLOCKSIZEY;
-    memsize_lbm = domainSize*sizeof(float)*9;
-    memsize_int = domainSize*sizeof(int);
-    memsize_float = domainSize*sizeof(float);
-    memsize_inputs = sizeof(m_obst_h);
-
-    float* fA_h = new float[domainSize * 9];
-    float* fB_h = new float[domainSize * 9];
-    float* floor_h = new float[domainSize];
-    int* im_h = new int[domainSize];
-
-    cudaMalloc((void **)&m_fA_d, memsize_lbm);
-    cudaMalloc((void **)&m_fB_d, memsize_lbm);
-    cudaMalloc((void **)&m_FloorTemp_d, memsize_float);
-    cudaMalloc((void **)&m_Im_d, memsize_int);
-    cudaMalloc((void **)&m_obst_d, memsize_inputs);
-}
-
-void CudaLbm::DeallocateDeviceMemory()
-{
-    cudaFree(m_fA_d);
-    cudaFree(m_fB_d);
-    cudaFree(m_Im_d);
-    cudaFree(m_FloorTemp_d);
-    cudaFree(m_obst_d);
-}
-
-void CudaLbm::InitializeDeviceMemory()
-{
-    int domainSize = ceil(MAX_XDIM / BLOCKSIZEX)*BLOCKSIZEX*ceil(MAX_YDIM / BLOCKSIZEY)*BLOCKSIZEY;
-    size_t memsize_lbm, memsize_float, memsize_inputs;
-    memsize_lbm = domainSize*sizeof(float)*9;
-    memsize_float = domainSize*sizeof(float);
-
-    float* f_h = new float[domainSize*9];
-    for (int i = 0; i < domainSize * 9; i++)
-    {
-        f_h[i] = 0;
-    }
-    cudaMemcpy(m_fA_d, f_h, memsize_lbm, cudaMemcpyHostToDevice);
-    cudaMemcpy(m_fB_d, f_h, memsize_lbm, cudaMemcpyHostToDevice);
-    delete[] f_h;
-    float* floor_h = new float[domainSize];
-    for (int i = 0; i < domainSize; i++)
-    {
-        floor_h[i] = 0;
-    }
-    cudaMemcpy(m_FloorTemp_d, floor_h, memsize_float, cudaMemcpyHostToDevice);
-    delete[] floor_h;
-
-    UpdateDeviceImage();
-
-    for (int i = 0; i < MAXOBSTS; i++)
-    {
-        m_obst_h[i].r1 = 0;
-        m_obst_h[i].x = 0;
-        m_obst_h[i].y = -1000;
-        m_obst_h[i].state = State::REMOVED;
-    }	
-    m_obst_h[0].r1 = 6.5;
-    m_obst_h[0].x = 30;// g_xDim*0.2f;
-    m_obst_h[0].y = 42;// g_yDim*0.3f;
-    m_obst_h[0].u = 0;// g_yDim*0.3f;
-    m_obst_h[0].v = 0;// g_yDim*0.3f;
-    m_obst_h[0].shape = Shape::SQUARE;
-    m_obst_h[0].state = State::NEW;
-
-    m_obst_h[1].r1 = 4.5;
-    m_obst_h[1].x = 30;// g_xDim*0.2f;
-    m_obst_h[1].y = 100;// g_yDim*0.3f;
-    m_obst_h[1].u = 0;// g_yDim*0.3f;
-    m_obst_h[1].v = 0;// g_yDim*0.3f;
-    m_obst_h[1].shape = Shape::VERTICAL_LINE;
-    m_obst_h[1].state = State::NEW;
-
-    memsize_inputs = sizeof(m_obst_h);
-    cudaMemcpy(m_obst_d, m_obst_h, memsize_inputs, cudaMemcpyHostToDevice);
-}
-
-void CudaLbm::UpdateDeviceImage()
-{
-    int domainSize = ceil(MAX_XDIM / BLOCKSIZEX)*BLOCKSIZEX*ceil(MAX_YDIM / BLOCKSIZEY)*BLOCKSIZEY;
-    int* im_h = new int[domainSize];
-    for (int i = 0; i < domainSize; i++)
-    {
-        int x = i%MAX_XDIM;
-        int y = i/MAX_XDIM;
-        im_h[i] = ImageFcn(x, y);
-    }
-    size_t memsize_int = domainSize*sizeof(int);
-    cudaMemcpy(m_Im_d, im_h, memsize_int, cudaMemcpyHostToDevice);
-    delete[] im_h;
-}
-
-int CudaLbm::ImageFcn(const int x, const int y){
-    int xDim = GetDomain()->GetXDim();
-    int yDim = GetDomain()->GetYDim();
-    if (x < 0.1f)
-        return 3;//west
-    else if ((xDim - x) < 1.1f)
-        return 2;//east
-    else if ((yDim - y) < 1.1f)
-        return 11;//11;//xsymmetry top
-    else if (y < 0.1f)
-        return 12;//12;//xsymmetry bottom
-    return 0;
-}
-
-Graphics::Graphics()
-{
-    m_shaderProgram = new ShaderProgram;
-    m_lightingProgram = new ShaderProgram;
-    m_obstProgram = new ShaderProgram;
-}
-
-void Graphics::CreateCudaLbm()
-{
-    m_cudaLbm = new CudaLbm;
-}
-
-CudaLbm* Graphics::GetCudaLbm()
-{
-    return m_cudaLbm;
-}
-
-cudaGraphicsResource* Graphics::GetCudaSolutionGraphicsResource()
-{
-    return m_cudaGraphicsResource;
-}
-
-void Graphics::CreateVboForCudaInterop(const unsigned int size)
-{
-    cudaGLSetGLDevice(gpuGetMaxGflopsDeviceId());
-    CreateVbo(size, cudaGraphicsMapFlagsWriteDiscard);
-    CreateElementArrayBuffer();
-}
-
-void Graphics::CleanUpGLInterOp()
-{
-    DeleteVbo();
-    DeleteElementArrayBuffer();
-}
-
-void Graphics::CreateVbo(const unsigned int size, const unsigned int vboResFlags)
-{
-    glGenVertexArrays(1, &m_vao);
-    glBindVertexArray(m_vao);
-
-    glGenBuffers(1, &m_vbo);
-    glBindBuffer(GL_ARRAY_BUFFER, m_vbo);
-
-    glBufferData(GL_ARRAY_BUFFER, size, 0, GL_DYNAMIC_DRAW);
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-    glBindVertexArray(0);
-
-    cudaGraphicsGLRegisterBuffer(&m_cudaGraphicsResource, m_vbo, vboResFlags);
-}
-
-void Graphics::DeleteVbo()
-{
-    cudaGraphicsUnregisterResource(m_cudaGraphicsResource);
-    glBindBuffer(1, m_vbo);
-    glDeleteBuffers(1, &m_vbo);
-    glDeleteVertexArrays(1, &m_vao);
-    m_vbo = 0;
-    m_vao = 0;
-}
-
-void Graphics::CreateElementArrayBuffer()
-{
-    const int numberOfElements = (MAX_XDIM - 1)*(MAX_YDIM - 1);
-    const int numberOfNodes = MAX_XDIM*MAX_YDIM;
-    GLuint* elementIndices = new GLuint[numberOfElements * 4 * 2];
-    for (int j = 0; j < MAX_YDIM-1; j++){
-        for (int i = 0; i < MAX_XDIM-1; i++){
-            //going clockwise, since y orientation will be flipped when rendered
-            elementIndices[j*(MAX_XDIM-1)*4+i*4+0] = (i)+(j)*MAX_XDIM;
-            elementIndices[j*(MAX_XDIM-1)*4+i*4+1] = (i+1)+(j)*MAX_XDIM;
-            elementIndices[j*(MAX_XDIM-1)*4+i*4+2] = (i+1)+(j+1)*MAX_XDIM;
-            elementIndices[j*(MAX_XDIM-1)*4+i*4+3] = (i)+(j+1)*MAX_XDIM;
-        }
-    }
-    for (int j = 0; j < MAX_YDIM-1; j++){
-        for (int i = 0; i < MAX_XDIM-1; i++){
-            //going clockwise, since y orientation will be flipped when rendered
-            elementIndices[numberOfElements*4+j*(MAX_XDIM-1)*4+i*4+0] = numberOfNodes+(i)+(j)*MAX_XDIM;
-            elementIndices[numberOfElements*4+j*(MAX_XDIM-1)*4+i*4+1] = numberOfNodes+(i+1)+(j)*MAX_XDIM;
-            elementIndices[numberOfElements*4+j*(MAX_XDIM-1)*4+i*4+2] = numberOfNodes+(i+1)+(j+1)*MAX_XDIM;
-            elementIndices[numberOfElements*4+j*(MAX_XDIM-1)*4+i*4+3] = numberOfNodes+(i)+(j+1)*MAX_XDIM;
-        }
-    }
-    glGenBuffers(1, &m_elementArrayBuffer);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_elementArrayBuffer);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(GLuint)*numberOfElements*4*2, elementIndices, GL_DYNAMIC_DRAW);
-    free(elementIndices);
-}
-
-void Graphics::DeleteElementArrayBuffer(){
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
-    glDeleteBuffers(1, &m_elementArrayBuffer);
-}
-
-template <typename T>
-void Graphics::CreateShaderStorageBuffer(T defaultValue, const unsigned int numberOfElements, const std::string name)
-{
-    GLuint temp;
-    glGenBuffers(1, &temp);
-    T* data = new T[numberOfElements];
-    for (int i = 0; i < numberOfElements; i++)
-    {
-        data[i] = defaultValue;
-    }
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, temp);
-    glBufferData(GL_SHADER_STORAGE_BUFFER, numberOfElements*sizeof(T), data, GL_STATIC_DRAW);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, temp);
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-    m_ssbos.push_back(Ssbo{ temp, name });
-}
-
-GLuint Graphics::GetShaderStorageBuffer(const std::string name)
-{
-    for (std::vector<Ssbo>::iterator it = m_ssbos.begin(); it != m_ssbos.end(); ++it)
-    {
-        if (it->m_name == name)
-        {
-            return it->m_id;
-        }
-    }
-}
-
-GLuint Graphics::GetElementArrayBuffer()
-{
-    return m_elementArrayBuffer;
-}
-
-GLuint Graphics::GetVbo()
-{
-    return m_vbo;
-}
-
-ShaderProgram* Graphics::GetShaderProgram()
-{
-    return m_shaderProgram;
-}
-
-ShaderProgram* Graphics::GetLightingProgram()
-{
-    return m_lightingProgram;
-}
-
-ShaderProgram* Graphics::GetObstProgram()
-{
-    return m_obstProgram;
-}
-
-void Graphics::CompileShaders()
-{
-    GetShaderProgram()->Initialize();
-    GetShaderProgram()->CreateShader("VertexShader.glsl", GL_VERTEX_SHADER);
-    GetShaderProgram()->CreateShader("FragmentShader.glsl", GL_FRAGMENT_SHADER);
-    GetLightingProgram()->Initialize();
-    GetLightingProgram()->CreateShader("ComputeShader.glsl", GL_COMPUTE_SHADER);
-    GetObstProgram()->Initialize();
-    GetObstProgram()->CreateShader("Obstructions.glsl", GL_COMPUTE_SHADER);
-}
-
-void Graphics::AllocateStorageBuffers()
-{
-    CreateShaderStorageBuffer(GLfloat(0), MAX_XDIM*MAX_YDIM*9, "LbmA");
-    CreateShaderStorageBuffer(GLfloat(0), MAX_XDIM*MAX_YDIM*9, "LbmB");
-    CreateShaderStorageBuffer(GLint(0), MAX_XDIM*MAX_YDIM, "Floor");
-    CreateShaderStorageBuffer(Obstruction{}, MAXOBSTS, "Obstructions");
-    CreateShaderStorageBuffer(float4{0,0,0,1e6}, 1, "RayIntersection");
-}
-
-void Graphics::InitializeObstSsbo()
-{
-    const GLuint obstSsbo = GetShaderStorageBuffer("Obstructions");
-    CudaLbm* cudaLbm = GetCudaLbm();
-    Obstruction* obst_h = cudaLbm->GetHostObst();
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, obstSsbo);
-    glBufferData(GL_SHADER_STORAGE_BUFFER, MAXOBSTS*sizeof(Obstruction), obst_h, GL_STATIC_DRAW);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, obstSsbo);
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-}
-
-int Graphics::RayCastMouseClick(float3 &rayCastIntersection, const float3 rayOrigin,
-    const float3 rayDir)
-{
-    Domain domain = *m_cudaLbm->GetDomain();
-    int xDim = domain.GetXDim();
-    int yDim = domain.GetYDim();
-    glm::vec4 intersectionCoord{ 0, 0, 0, 0 };
-    float intersectCoord[4];
-    const GLuint ssbo_rayIntersection = GetShaderStorageBuffer("RayIntersection");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 4, ssbo_rayIntersection);
-
-    ShaderProgram* const shader = GetLightingProgram();
-    shader->Use();
-
-    GLuint shaderID = shader->GetId();
-    SetUniform(shaderID, "maxXDim", MAX_XDIM);
-    SetUniform(shaderID, "maxyDim", MAX_YDIM);
-    SetUniform(shaderID, "maxObsts", MAXOBSTS);
-    SetUniform(shaderID, "xDim", xDim);
-    SetUniform(shaderID, "yDim", yDim);
-    SetUniform(shaderID, "xDimVisible", domain.GetXDimVisible());
-    SetUniform(shaderID, "yDimVisible", domain.GetYDimVisible());
-    SetUniform(shaderID, "rayOrigin", rayOrigin);
-    SetUniform(shaderID, "rayDir", rayDir);
-
-    RunSubroutine(shaderID, "ResetRayCastData", int3{ 1, 1, 1 });
-    RunSubroutine(shaderID, "RayCast", int3{ xDim, yDim, 1 });
-
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, ssbo_rayIntersection);
-    GLfloat* intersect = (GLfloat*)glMapBufferRange(GL_SHADER_STORAGE_BUFFER, 0,
-        sizeof(glm::vec4), GL_MAP_READ_BIT);
-    std::memcpy(&intersectionCoord, intersect, sizeof(glm::vec4));
-    glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
-
-    int returnVal;
-    if (intersectionCoord.w > 1e5) //ray did not intersect with any objects
-    {
-        returnVal = 1;
-    }
-    else
-    {
-        RunSubroutine(shaderID, "ResetRayCastData", int3{ 1, 1, 1 });
-        rayCastIntersection.x = intersectionCoord.x;
-        rayCastIntersection.y = intersectionCoord.y;
-        rayCastIntersection.z = intersectionCoord.z;
-        returnVal = 0;
-    }
-    shader->Unset();
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-    return returnVal;
-}
-
-void Graphics::RenderVbo(const bool renderFloor, Domain &domain, const glm::mat4 &modelMatrix,
-    const glm::mat4 &projectionMatrix)
-{
-    glEnable(GL_BLEND);
-    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-
-    glMatrixMode(GL_MODELVIEW);
-    glLoadIdentity();
-
-    glBindVertexArray(m_vao);
-    //Draw solution field
-    glBindBuffer(GL_ARRAY_BUFFER, m_vbo);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_elementArrayBuffer);
-    glEnableVertexAttribArray(0);
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 16, 0);
-    glEnableVertexAttribArray(1);
-    glVertexAttribPointer(1, 1, GL_FLOAT, GL_FALSE, 16, (GLvoid*)(3 * sizeof(GLfloat)));
-    glEnableClientState(GL_VERTEX_ARRAY);
-
-    int yDimVisible = domain.GetYDimVisible();
-    if (renderFloor)
-    {
-        //Draw floor
-        glDrawElements(GL_QUADS, (MAX_XDIM - 1)*(yDimVisible - 1)*4, GL_UNSIGNED_INT, 
-            BUFFER_OFFSET(sizeof(GLuint)*4*(MAX_XDIM - 1)*(MAX_YDIM - 1)));
-    }
-    //Draw water surface
-    glDrawElements(GL_QUADS, (MAX_XDIM - 1)*(yDimVisible - 1)*4 , GL_UNSIGNED_INT, (GLvoid*)0);
-    glDisableClientState(GL_VERTEX_ARRAY);
-    glBindVertexArray(0);
-}
-
-void Graphics::RunComputeShader(const float3 cameraPosition)
-{
-    const GLuint ssbo_lbmA = GetShaderStorageBuffer("LbmA");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, ssbo_lbmA);
-    const GLuint ssbo_lbmB = GetShaderStorageBuffer("LbmB");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 1, ssbo_lbmB);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 2, m_vbo);
-    const GLuint ssbo_floor = GetShaderStorageBuffer("Floor");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 3, ssbo_floor);
-    const GLuint ssbo_obsts = GetShaderStorageBuffer("Obstructions");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 5, ssbo_obsts);
-    ShaderProgram* const shader = GetLightingProgram();
-
-    shader->Use();
-
-    Domain domain = *m_cudaLbm->GetDomain();
-    const int xDim = domain.GetXDim();
-    const int yDim = domain.GetYDim();
-    GLuint shaderID = shader->GetId();
-    SetUniform(shaderID, "maxXDim", MAX_XDIM);
-    SetUniform(shaderID, "maxyDim", MAX_YDIM);
-    SetUniform(shaderID, "maxObsts", MAXOBSTS);
-    SetUniform(shaderID, "xDim", xDim);
-    SetUniform(shaderID, "yDim", yDim);
-    SetUniform(shaderID, "xDimVisible", domain.GetXDimVisible());
-    SetUniform(shaderID, "yDimVisible", domain.GetYDimVisible());
-    SetUniform(shaderID, "cameraPosition", cameraPosition);
-    SetUniform(shaderID, "uMax", m_inletVelocity);
-    SetUniform(shaderID, "omega", m_omega);
-
-    for (int i = 0; i < 5; i++)
-    {
-        RunSubroutine(shaderID, "MarchLbm", int3{ xDim, yDim, 1 });
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 1, ssbo_lbmA);
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, ssbo_lbmB);
-
-        RunSubroutine(shaderID, "MarchLbm", int3{ xDim, yDim, 1 });
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, ssbo_lbmA);
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 1, ssbo_lbmB);
-    }
-    
-    RunSubroutine(shaderID, "UpdateFluidVbo", int3{ xDim, yDim, 1 });
-    RunSubroutine(shaderID, "DeformFloorMeshUsingCausticRay", int3{ xDim, yDim, 1 });
-    RunSubroutine(shaderID, "ComputeFloorLightIntensitiesFromMeshDeformation", int3{ xDim, yDim, 1 });
-    RunSubroutine(shaderID, "ApplyCausticLightingToFloor", int3{ xDim, yDim, 1 });
-    RunSubroutine(shaderID, "PhongLighting", int3{ xDim, yDim, 2 });
-    RunSubroutine(shaderID, "UpdateObstructionTransientStates", int3{ xDim, yDim, 1 });
-    RunSubroutine(shaderID, "CleanUpVbo", int3{ MAX_XDIM, MAX_YDIM, 2 });
-    
-    shader->Unset();
-
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-}
-
-
-
-void Graphics::UpdateObstructionsUsingComputeShader(const int obstId, Obstruction &newObst)
-{
-    const GLuint ssbo_obsts = GetShaderStorageBuffer("Obstructions");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, ssbo_obsts);
-    ShaderProgram* const shader = GetObstProgram();
-    shader->Use();
-
-    GLuint shaderId = shader->GetId();
-
-    SetUniform(shaderId, "targetObst.shape", newObst.shape);
-    SetUniform(shaderId, "targetObst.x", newObst.x);
-    SetUniform(shaderId, "targetObst.y", newObst.y);
-    SetUniform(shaderId, "targetObst.r1", newObst.r1);
-    SetUniform(shaderId, "targetObst.u", newObst.u);
-    SetUniform(shaderId, "targetObst.v", newObst.v);
-    SetUniform(shaderId, "targetObst.state", newObst.state);
-    SetUniform(shaderId, "targetObstId", obstId);
-
-    RunSubroutine(shader->GetId(), "UpdateObstruction", int3{ 1, 1, 1 });
-
-    shader->Unset();
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-}
-
-void Graphics::InitializeComputeShaderData()
-{
-    const GLuint ssbo_lbmA = GetShaderStorageBuffer("LbmA");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, ssbo_lbmA);
-    const GLuint ssbo_lbmB = GetShaderStorageBuffer("LbmB");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 1, ssbo_lbmB);
-    const GLuint ssbo_obsts = GetShaderStorageBuffer("Obstructions");
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 5, ssbo_obsts);
-    ShaderProgram* const shader = GetLightingProgram();
-
-    shader->Use();
-
-    Domain domain = *m_cudaLbm->GetDomain();
-    GLuint shaderId = shader->GetId();
-    SetUniform(shaderId, "maxXDim", MAX_XDIM);
-    SetUniform(shaderId, "maxYDim", MAX_YDIM);
-    SetUniform(shaderId, "maxObsts", MAXOBSTS);//
-    SetUniform(shaderId, "xDim", domain.GetXDim());
-    SetUniform(shaderId, "yDim", domain.GetYDim());
-    SetUniform(shaderId, "xDimVisible", domain.GetXDim());
-    SetUniform(shaderId, "yDimVisible", domain.GetYDim());
-    SetUniform(shaderId, "uMax", m_inletVelocity);
-
-    RunSubroutine(shaderId, "InitializeDomain", int3{ MAX_XDIM, MAX_YDIM, 1 });
-
-    shader->Unset();
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-}
-    
-void Graphics::SetOmega(const float omega)
-{
-    m_omega = omega;
-}
-
-float Graphics::GetOmega()
-{
-    return m_omega;
-}
-
-void Graphics::SetInletVelocity(const float u)
-{
-    m_inletVelocity = u;
-}
-
-float Graphics::GetInletVelocity()
-{
-    return m_inletVelocity;
-}
-
-void Graphics::UpdateLbmInputs(const float u, const float omega)
-{
-    SetInletVelocity(u);
-    SetOmega(omega);
-}
-
-void Graphics::RenderVboUsingShaders(const bool renderFloor, Domain &domain,
-    const glm::mat4 &modelMatrix, const glm::mat4 &projectionMatrix)
-{
-    ShaderProgram* shader = GetShaderProgram();
-
-    shader->Use();//
-    GLint modelMatrixLocation = glGetUniformLocation(shader->GetId(), "modelMatrix");
-    GLint projectionMatrixLocation = glGetUniformLocation(shader->GetId(), "projectionMatrix");
-    glUniformMatrix4fv(modelMatrixLocation, 1, GL_FALSE, glm::value_ptr(modelMatrix));
-    glUniformMatrix4fv(projectionMatrixLocation, 1, GL_FALSE, glm::value_ptr(projectionMatrix));
-
-    RenderVbo(renderFloor, domain, modelMatrix, projectionMatrix);
-
-    shader->Unset();
-    glBindVertexArray(0);
-}
-
-
 GraphicsManager::GraphicsManager(Panel* panel)
 {
     m_parent = panel;
-    m_graphics = new Graphics;
+    m_graphics = new ShaderManager;
     m_graphics->CreateCudaLbm();
     m_obstructions = m_graphics->GetCudaLbm()->GetHostObst();
     m_rotate = { 60.f, 0.f, 30.f };
@@ -734,7 +117,7 @@ CudaLbm* GraphicsManager::GetCudaLbm()
     return m_graphics->GetCudaLbm();
 }
 
-Graphics* GraphicsManager::GetGraphics()
+ShaderManager* GraphicsManager::GetGraphics()
 {
     return m_graphics;
 }
@@ -786,13 +169,13 @@ void GraphicsManager::SetUpGLInterop()
 {
     unsigned int solutionMemorySize = MAX_XDIM*MAX_YDIM * 4 * sizeof(float);
     unsigned int floorSize = MAX_XDIM*MAX_YDIM * 4 * sizeof(float);
-    Graphics* graphics = GetGraphics();
+    ShaderManager* graphics = GetGraphics();
     graphics->CreateVboForCudaInterop(solutionMemorySize+floorSize);
 }
 
 void GraphicsManager::SetUpShaders()
 {
-    Graphics* graphics = GetGraphics();
+    ShaderManager* graphics = GetGraphics();
     graphics->CompileShaders();
     graphics->AllocateStorageBuffers();
     graphics->InitializeObstSsbo();
@@ -819,7 +202,7 @@ void GraphicsManager::SetUpCuda()
     float* floor_d = cudaLbm->GetFloorTemp();
     Obstruction* obst_d = cudaLbm->GetDeviceObst();
 
-    Graphics* graphics = GetGraphics();
+    ShaderManager* graphics = GetGraphics();
     cudaGraphicsResource* cudaSolutionField = graphics->GetCudaSolutionGraphicsResource();
     float4 *dptr;
     cudaGraphicsMapResources(1, &cudaSolutionField, 0);
@@ -839,7 +222,7 @@ void GraphicsManager::RunCuda()
 {
     // map OpenGL buffer object for writing from CUDA
     CudaLbm* cudaLbm = GetCudaLbm();
-    Graphics* graphics = GetGraphics();
+    ShaderManager* graphics = GetGraphics();
     cudaGraphicsResource* vbo_resource = graphics->GetCudaSolutionGraphicsResource();
     Panel* rootPanel = m_parent->GetRootPanel();
 
@@ -1268,7 +651,7 @@ void GraphicsManager::UpdateLbmInputs()
     CudaLbm* cudaLbm = GetCudaLbm();
     cudaLbm->SetInletVelocity(u);
     cudaLbm->SetOmega(omega);
-    Graphics* graphics = GetGraphics();
+    ShaderManager* graphics = GetGraphics();
     graphics->UpdateLbmInputs(u, omega);
 }
 
@@ -1297,36 +680,3 @@ float GetDistanceBetweenTwoPoints(const float x1, const float y1,
     return sqrt(dx*dx + dy*dy);
 }
 
-
-void SetUniform(GLuint shaderId, const GLchar* varName, const int varValue)
-{
-    const GLint targetLocation = glGetUniformLocation(shaderId, varName);
-    glUniform1i(targetLocation, varValue);
-}
-
-void SetUniform(GLuint shaderId, const GLchar* varName, const float varValue)
-{
-    const GLint targetLocation = glGetUniformLocation(shaderId, varName);
-    glUniform1f(targetLocation, varValue);
-}
-
-void SetUniform(GLuint shaderId, const GLchar* varName, const bool varValue)
-{
-    const GLint targetLocation = glGetUniformLocation(shaderId, varName);
-    glUniform1i(targetLocation, varValue);
-}
-
-void SetUniform(GLuint shaderId, const GLchar* varName, const float3 varValue)
-{
-    const GLint targetLocation = glGetUniformLocation(shaderId, varName);
-    glUniform3f(targetLocation, varValue.x, varValue.y, varValue.z);
-}
-
-void RunSubroutine(GLuint shaderId, const GLchar* subroutineName, const int3 workGroupSize)
-{
-    const GLuint subroutine = glGetSubroutineIndex(shaderId, GL_COMPUTE_SHADER,
-        subroutineName);
-    glUniformSubroutinesuiv(GL_COMPUTE_SHADER, 1, &subroutine);
-    glDispatchCompute(workGroupSize.x, workGroupSize.y, workGroupSize.z);
-    glMemoryBarrier(GL_SHADER_STORAGE_BARRIER_BIT);
-}
