+++ InteractiveCfd_Core/Graphics/GraphicsManager.h
@@ -2,14 +2,10 @@
 #include "common.h"
 #include "cuda_runtime.h"
 #include <GLEW/glew.h>
-#include "cuda_gl_interop.h"  // needs GLEW
-#include "helper_cuda.h"
-#include "helper_cuda_gl.h"
 #include <glm/glm.hpp>
 #include <vector>
 #include <string>
 
-#define BUFFER_OFFSET(i) ((char *)NULL + (i))
 
 #ifdef LBM_GL_CPP_EXPORTS  
 #define FW_API __declspec(dllexport)   
@@ -18,115 +14,8 @@
 #endif  
 
 class Panel;
-class Mouse;
-class Domain;
-class ShaderProgram;
-
-class FW_API CudaLbm
-{
-private:
-    int m_maxX;
-    int m_maxY;
-    Domain* m_domain;
-    float* m_fA_d;
-    float* m_fB_d;
-    int* m_Im_d;
-    float* m_FloorTemp_d;
-    Obstruction* m_obst_d;
-    Obstruction m_obst_h[MAXOBSTS];
-    float m_inletVelocity;
-    float m_omega;
-    bool m_isPaused;
-    int m_timeStepsPerFrame;
-public:
-    CudaLbm();
-    CudaLbm(const int maxX, const int maxY);
-    Domain* GetDomain();
-    float* GetFA();
-    float* GetFB();
-    int* GetImage();
-    float* GetFloorTemp();
-    Obstruction* GetDeviceObst();
-    Obstruction* GetHostObst();
-    float GetInletVelocity();
-    float GetOmega();
-    void SetInletVelocity(const float velocity);
-    void SetOmega(const float omega);
-    void SetPausedState(const bool isPaused);
-    bool IsPaused();
-    int GetTimeStepsPerFrame();
-    void SetTimeStepsPerFrame(const int timeSteps);
-
-    void AllocateDeviceMemory();
-    void InitializeDeviceMemory();
-    void DeallocateDeviceMemory();
-    void UpdateDeviceImage();
-    int ImageFcn(const int x, const int y);
-
-   
-};
-
-class FW_API Graphics
-{
-private:
-    class Ssbo
-    {
-    public:
-        GLuint m_id;
-        std::string m_name;
-    };
-    CudaLbm* m_cudaLbm;
-    cudaGraphicsResource* m_cudaGraphicsResource;
-    GLuint m_vao;
-    GLuint m_vbo;
-    GLuint m_elementArrayBuffer;
-    ShaderProgram* m_shaderProgram;
-    ShaderProgram* m_lightingProgram;
-    ShaderProgram* m_obstProgram;
-    std::vector<Ssbo> m_ssbos;
-    float m_omega;
-    float m_inletVelocity;
-public:
-    Graphics();
-
-    void CreateCudaLbm();
-    CudaLbm* GetCudaLbm();
-    cudaGraphicsResource* GetCudaSolutionGraphicsResource();
-    GLuint GetVbo();
-    GLuint GetElementArrayBuffer();
-    void CreateVbo(const unsigned int size, const unsigned int vboResFlags);
-    void DeleteVbo();
-    void CreateElementArrayBuffer();
-    void DeleteElementArrayBuffer();
-    template <typename T> void CreateShaderStorageBuffer(T defaultValue,
-        const unsigned int sizeInInts, const std::string name);
-    GLuint GetShaderStorageBuffer(const std::string name);
-    void CreateVboForCudaInterop(unsigned int size);
-    void CleanUpGLInterOp();
-    ShaderProgram* GetShaderProgram();
-    ShaderProgram* GetLightingProgram();
-    ShaderProgram* GetObstProgram();
-    void CompileShaders();
-    void AllocateStorageBuffers();
-    void InitializeObstSsbo();
-    void InitializeComputeShaderData();
-    
-    void SetOmega(const float omega);
-    float GetOmega();
-    void SetInletVelocity(const float u);
-    float GetInletVelocity();
-    void UpdateLbmInputs(const float u, const float omega);
-    void RunComputeShader(const float3 cameraPosition);
-    void UpdateObstructionsUsingComputeShader(const int obstId, Obstruction &newObst);
-    int RayCastMouseClick(float3 &rayCastIntersection, const float3 rayOrigin,
-        const float3 rayDir);
-    void RenderVbo(const bool renderFloor, Domain &domain, const glm::mat4 &modelMatrix,
-        const glm::mat4 &projectionMatrix);
-    void RenderVboUsingShaders(const bool renderFloor, Domain &domain,
-        const glm::mat4 &modelMatrix, const glm::mat4 &projectionMatrix);
-};
-
-
+class ShaderManager;
+class CudaLbm;
 
 class FW_API GraphicsManager
 {
@@ -148,7 +37,7 @@ private:
     float m_contourMinValue;
     float m_contourMaxValue;
     ContourVariable m_contourVar;
-    Graphics* m_graphics;
+    ShaderManager* m_graphics;
     bool m_useCuda = true;
 
 public:
@@ -182,7 +71,7 @@ public:
     void SetScaleFactor(const float scaleFactor);
 
     CudaLbm* GetCudaLbm();
-    Graphics* GetGraphics();
+    ShaderManager* GetGraphics();
 
     void CenterGraphicsViewToGraphicsPanel(const int leftPanelWidth);
     void SetUpGLInterop();
@@ -233,9 +122,3 @@ public:
 
 float GetDistanceBetweenTwoPoints(const float x1, const float y1,
     const float x2, const float y2);
-void SetUniform(GLuint shaderId, const GLchar* varName, const int varValue);
-void SetUniform(GLuint shaderId, const GLchar* varName, const float varValue);
-void SetUniform(GLuint shaderId, const GLchar* varName, const bool varValue);
-void SetUniform(GLuint shaderId, const GLchar* varName, const float3 varValue);
-void RunSubroutine(GLuint shaderId, const GLchar* subroutineName,
-    const int3 workGroupSize);
