+++ InteractiveCfd_GlutWindow/Window.cpp
@@ -1,7 +1,8 @@
 #include "Window.h"
 #include "Layout.h"
 #include "Panel.h"
-#include "GraphicsManager.h"
+#include "Graphics/GraphicsManager.h"
+#include "Graphics/CudaLbm.h"
 #include "Domain.h"
 #include <GLUT/freeglut.h>
 #include <typeinfo>
