+++ InteractiveCfd_GlutWindow/main.cpp
@@ -1,7 +1,7 @@
 #include "Window.h"
 #include "Layout.h"
 #include "Panel.h"
-#include "GraphicsManager.h"
+#include "Graphics/GraphicsManager.h"
 #include <GLEW/glew.h>
 #include <GLUT/freeglut.h>
